chore: fix typings, update test outputs, and add requirements.txt
- Fixed type annotations for SUMO/TraCI functions to resolve Pylance errors
- Generated new simulation outputs from testing for validation
- Added requirements.txt to capture dependencies for reproducibility

+++ src/algorithms/websters/websters.py
@@ -22,7 +22,7 @@ def websters_method(
     Returns:
         int: Optimal Cycle Length (s)
     """
-    return ceil(1.5 * L + 5) / (1 - Y)
+    return int(ceil(1.5 * L + 5) / (1 - Y))
 
 
 def compute_signal_config_with_poisson(
